+++ changelog.md
@@ -36,6 +36,7 @@
     - `bytm:dataStoreSerializerLoaded` - Emitted after all memory-cached DataStore instances' data was lazy-loaded.
   - All DataStore instances that have in-memory cached data will now be lazy-loaded after feature initialization is done. Note: lazy-loading starts after `bytm:allReady`, so it's more aimed at lowering initial data access times.
   - Features can now have tags associated with them, which is another way of filtering them, like when using the new internal functions `configSetFeatsWithTags()` and `getFeaturesWithTags()`.
+  - Added an advanced-mode feature that allows SelectorObserver checks and found elements to be logged to the console for performance debugging.
   - Updated the dependencies CoreUtils to v3.8.0 and UserUtils to v11.0.0
   - The `BytmDialog` and `MarkdownDialog` render functions now receive the dialog instance as the sole parameter.
   - Changed the ID of the `BytmDialog` returned by `showPrompt()` from `prompt-dialog` to just `prompt`